backend/requirements.txt
asgiref==3.11.1
Django==5.0.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.5
psycopg2-binary>=2.9
drf-spectacular>=0.28
gunicorn
